Move CChancesDrawConfiguration traits
